[ADC] Add Touch/Hall Effect for ESP32 + ADC pin selection
Can be configured to send an event when an ADC pin was touched.

+++ src/ESPEasy-Globals.cpp
@@ -49,7 +49,7 @@ unsigned long last_system_event_run = 0;
 #if FEATURE_ADC_VCC
 float vcc = -1.0;
 #endif
-uint16_t lastADCvalue = 0;
+int lastADCvalue = 0;
 
 boolean WebLoggedIn = false;
 int WebLoggedInTimer = 300;
