+++ src/ESPEasy-Globals.h
@@ -313,7 +313,7 @@ extern unsigned long last_system_event_run;
 extern float vcc;
 #endif
 #ifdef ESP8266
-extern uint16_t lastADCvalue; // Keep track of last ADC value as it cannot be read while WiFi is connecting
+extern int lastADCvalue; // Keep track of last ADC value as it cannot be read while WiFi is connecting
 #endif
 
 extern boolean WebLoggedIn;
